refactors interface
