+++ squidCore/src/main/java/org/cirdles/squid/tasks/TaskInterface.java
@@ -19,7 +19,6 @@ import org.cirdles.squid.tasks.expressions.spots.SpotSummaryDetails;
 import com.thoughtworks.xstream.XStream;
 import java.util.List;
 import java.util.Map;
-import org.cirdles.squid.SquidDataFileInterface;
 import org.cirdles.squid.constants.Squid3Constants;
 import org.cirdles.squid.core.CalamariReportsEngine;
 import org.cirdles.squid.parameters.parameterModels.ParametersModel;
@@ -34,6 +33,7 @@ import org.cirdles.squid.shrimp.SquidSessionModel;
 import org.cirdles.squid.shrimp.SquidSpeciesModel;
 import org.cirdles.squid.tasks.expressions.Expression;
 import org.cirdles.squid.tasks.expressions.expressionTrees.ExpressionTreeInterface;
+import org.cirdles.squid.shrimp.ShrimpDataFileInterface;
 
 /**
  *
@@ -109,7 +109,7 @@ public interface TaskInterface {
     /**
      * @param prawnFile the prawnFile to set
      */
-    public void setPrawnFile(SquidDataFileInterface prawnFile);
+    public void setPrawnFile(ShrimpDataFileInterface prawnFile);
 
     public SquidSpeciesModel lookUpSpeciesByName(String isotopeName);
 
@@ -255,7 +255,7 @@ public interface TaskInterface {
      */
     public List<ShrimpFractionExpressionInterface> getShrimpFractions();
 
-    public List<ShrimpFractionExpressionInterface> processRunFractions(SquidDataFileInterface prawnFile, SquidSessionModel squidSessionSpecs);
+    public List<ShrimpFractionExpressionInterface> processRunFractions(ShrimpDataFileInterface prawnFile, SquidSessionModel squidSessionSpecs);
 
     /**
      * @return the referenceMaterialSpots
